\name{add} 
\alias{add} 
\title{ add two numbers } 
\description{ 
    A initial starting function which can add two numbers. 
} 
\usage{ add(x, y) } 
\arguments{ 
  \item{x}{ 
    numercial value 
  } 
  \item{y}{ 
    numerical value 
  } 
} 
\details{ 
    Some more details ... 
} 
\value{return the sum of x and} 
\examples{ 
    add(2,3) 
    x=2 
    y=3 
    add(x,y) 
} 
\seealso{  See also \code{\link{testprint}} } 
 
\keyword{ arith } 
