\name{testprint} 
\alias{testprint} 
\title{ print a test message } 
\description{ 
    A test function which prints a string. 
} 
\usage{ testprint(txt) } 
\arguments{ 
  \item{txt}{ some value to print  } 
} 
\examples{ 
    testprint("Hello World!") 
} 
\seealso{  See also \code{\link{add}} } 
\keyword{ arith } 
